解决冲突

        https://codeup.aliyun.com/fly/Link/Link_Service/merge_request/10146218

+++ app/user/restful/internal/logic/getuserinfologic.go
@@ -40,7 +40,7 @@ func (l *GetUserInfoLogic) GetUserInfo(req *types.UserInfoRequest) (resp *types.
 		Username:  response.Username,
 		Avatar:    response.Avatar,
 		Age:       uint(response.Age),
-		Gender:    response.Gender,
+		Gender:    uint(response.Gender),
 		Address:   response.Address,
 		Phone:     response.Phone,
 		Signature: response.Signature,
